% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statStudent.R
\docType{data}
\name{GeomStudent}
\alias{GeomStudent}
\title{GeomStudent}
\format{An object of class \code{GeomStudent} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\description{
GeomStudent
}
\keyword{internal}
