% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statStudent.R
\docType{data}
\name{StatStudent}
\alias{StatStudent}
\title{Bits and pieces copied from ggplot2 sources and
https://github.com/wjschne/ggnormalviolin}
\format{An object of class \code{StatStudent} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.}
\description{
StatStudent
}
\keyword{internal}
